+++ src/containers/App.js
@@ -5,6 +5,15 @@ import zhLocaleData from 'react-intl/locale-data/zh'
 // import locale data
 import { addLocaleData, IntlProvider } from 'react-intl'
 import { connect } from 'react-redux'
+import { signOutAction } from '@twreporter/registration'
+import Layout from '../components/Layout'
+
+// lodash
+import get from 'lodash/get'
+
+const _ = {
+  get
+}
 
 addLocaleData(enLocaleData)
 addLocaleData(zhLocaleData)
@@ -27,9 +36,17 @@ class App extends PureComponent {
   }
 
   render() {
+    const pathname = _.get(this.props, 'location.pathname')
     return (
       <IntlProvider locale={currentLocale} defaultLocale="zh-Hant">
-        {this.props.children}
+        <Layout
+          header={this.props.header}
+          pathname={pathname}
+          signOutAction={this.props.signOutAction}
+          ifAuthenticated={this.props.ifAuthenticated}
+        >
+          {this.props.children}
+        </Layout>
       </IntlProvider>
     )
   }
@@ -40,11 +57,21 @@ App.childContextTypes = {
   device: React.PropTypes.string
 }
 
+App.defaultProps = {
+  signOutAction: () => {},
+  ifAuthenticated: false
+}
+
+App.propTypes = {
+  signOutAction: React.PropTypes.func,
+  ifAuthenticated: React.PropTypes.bool
+}
+
 function mapStateToProps(state) {
   return {
     header: state.header,
-    device: state.device
+    ifAuthenticated: _.get(state, [ 'auth', 'authenticated' ], false)
   }
 }
 
-export default connect(mapStateToProps)(App)
+export default connect(mapStateToProps, { signOutAction })(App)